ingestor/requirements.txt
python-dotenv
redis
paho-mqtt
prometheus_client